Fixing bugs and adding base display
Fixed various issues with parsed instructions where the wrong element was added to the state and a number of problems concerning out of bounds exceptions when placing the elements in a randomised world. Issues still remaining are: pseudo number generator quickly exceeds the integer representation max size and wraps back around using negative numbers, GUI not implemented and so cannot be connected, Issues with placement/visualisation of structures dependant on the rows which they are placed. Shifting single places to the left or right in a consistent manner, need to check that adjacent cell function works correctly and that the ants can pickup food as the functionality did not appear in the simulation.